+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added job contexts
 - Added jobIDs for easier job tracking
 - Added output filesize to structured logging
+- Adding per package framework for varying log levels
 
 ## [0.92.24] - 2025-6-8
 - Deployed & implement multiple fields for more effective structured logging for the project long term